+++ review/src/main/java/com/ticketing/review/domain/event/ReviewDeleteErrorEvent.java
@@ -1,4 +1,4 @@
-package com.ticketing.review.application.dto.event;
+package com.ticketing.review.domain.event;
 
 
 import java.util.UUID;
